refactor

+++ src/services/orderService.js
@@ -2,11 +2,9 @@ import Order from "../models/Order.js";
 import Payment from "../models/Payment.js";
 import crypto from "crypto";
 import payment from "../utils/payment.js";
-import {
-  ORDER_STATUS_CONFIRMED,
-  ORDER_STATUS_PENDING,
-} from "../constants/orderStatuses.js";
+import { ORDER_STATUS_CONFIRMED } from "../constants/orderStatuses.js";
 import { PAYMENT_STATUS_COMPLETED } from "../constants/paymenStatuses.js";
+import { ADMIN } from "../constants/roles.js";
 
 const getOrders = async () => {
   const orders = await Order.find()
@@ -47,7 +45,16 @@ const createOrder = async (data, userId) => {
   return await Order.create({ ...data, user: userId, orderNumber });
 };
 
-const updateOrder = async (id, data) => {
+const updateOrder = async (id, data, user) => {
+  const order = await getOrderById(id);
+
+  if (order.user._id != user._id && !user.roles.includes(ADMIN)) {
+    throw {
+      statusCode: 403,
+      message: "Access denied.",
+    };
+  }
+
   return await Order.findByIdAndUpdate(
     id,
     {
@@ -57,12 +64,28 @@ const updateOrder = async (id, data) => {
   );
 };
 
-const deleteOrder = async (id) => {
+const deleteOrder = async (id, user) => {
+  const order = await getOrderById(id);
+
+  if (order.user._id != user._id && !user.roles.includes(ADMIN)) {
+    throw {
+      statusCode: 403,
+      message: "Access denied.",
+    };
+  }
+
   return await Order.findByIdAndDelete(id);
 };
 
-const orderPaymentViaKhalti = async (id) => {
+const orderPaymentViaKhalti = async (id, user) => {
   const order = await getOrderById(id);
+
+  if (order.user._id != user._id) {
+    throw {
+      statusCode: 403,
+      message: "Access denied.",
+    };
+  }
 
   const transactionId = crypto.randomUUID();
 
@@ -84,8 +107,15 @@ const orderPaymentViaKhalti = async (id) => {
   });
 };
 
-const confirmOrderPayment = async (id, status) => {
+const confirmOrderPayment = async (id, status, user) => {
   const order = await getOrderById(id);
+
+  if (order.user._id != user._id && !user.roles.includes(ADMIN)) {
+    throw {
+      statusCode: 403,
+      message: "Access denied.",
+    };
+  }
 
   if (status.toUpperCase() != PAYMENT_STATUS_COMPLETED) {
     await Payment.findByIdAndUpdate(order.payment._id, {
